refactor(Test): move BottomTabsContainer and StackContainer to shared components folder (#161)
## Description

This PR moves \*Container and LayoutView components to shared components
folder for use across different tests.

## Changes

- moved BottomTypesContainer and StackContainer + adjusted related
import statements
- moved LayoutView and renamed to CenteredLayoutView

## Testing

Nothing should be changed while running current test app.
