+++ apps/src/shared/gamma/containers/Stack/StackContainer.tsx
