+++ apps/src/tests/TestBottomTabs/tabs/Tab1.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { LayoutView } from '../components/LayoutView';
+import { CenteredLayoutView } from '../../../shared/CenteredLayoutView';
 import { TabContentView } from '../components/TabContentView';
 import Colors from '../../../shared/styling/Colors';
 
 export function Tab1() {
   return (
-    <LayoutView style={{ backgroundColor: Colors.OffWhite }}>
+    <CenteredLayoutView style={{ backgroundColor: Colors.OffWhite }}>
       <TabContentView selectNextTab={undefined} tabKey={'Tab1'} />
-    </LayoutView>
+    </CenteredLayoutView>
   );
 }
