+++ apps/src/tests/TestBottomTabs/tabs/Tab2.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { LayoutView } from '../components/LayoutView';
+import { CenteredLayoutView } from '../../../shared/CenteredLayoutView';
 import { TabContentView } from '../components/TabContentView';
 import Colors from '../../../shared/styling/Colors';
 
 export function Tab2() {
   return (
-    <LayoutView style={{ backgroundColor: Colors.PurpleLight100 }}>
+    <CenteredLayoutView style={{ backgroundColor: Colors.PurpleLight100 }}>
       <TabContentView selectNextTab={undefined} tabKey={'Tab2'} />
-    </LayoutView>
+    </CenteredLayoutView>
   );
 }
