+++ apps/src/tests/TestBottomTabs/tabs/Tab3.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import { LayoutView } from '../components/LayoutView';
+import { CenteredLayoutView } from '../../../shared/CenteredLayoutView';
 import { TabContentView } from '../components/TabContentView';
 import Colors from '../../../shared/styling/Colors';
 
 export function Tab3() {
   // const
   return (
-    <LayoutView style={{ backgroundColor: Colors.YellowLight100 }}>
+    <CenteredLayoutView style={{ backgroundColor: Colors.YellowLight100 }}>
       <TabContentView selectNextTab={undefined} tabKey={'Tab3'} />
-    </LayoutView>
+    </CenteredLayoutView>
   );
 }
