+++ apps/src/tests/TestScreenStack/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { generateStackWithNames } from './helper';
-import { StackContainer } from './StackContainer';
+import { StackContainer } from '../../shared/gamma/containers/stack/StackContainer';
 
 export default function App() {
   return (
